fix: function naming & position (sortChild => sortChildren)
