+++ packages/replies/src/utils.ts
@@ -9,16 +9,3 @@ export function checkUniqueReply(reply: Reply[]): Reply[] {
 
   return result
 }
-
-export function sortChild(reply: Reply): Reply {
-  const sortedChildReply = reply.children.sort(
-    (a, b) => new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime(),
-  )
-
-  const result = {
-    ...reply,
-    children: sortedChildReply,
-  }
-
-  return result
-}
